+++ lizard_annotation/forms.py
@@ -1,26 +1,20 @@
 from django import forms
 from django.utils.translation import ugettext
 
-from models import Annotation
 from models import AnnotationStatus
 from models import AnnotationCategory
 from models import AnnotationType
 
 
-annotation_status_choices = tuple(
-    (ans.status, ans.status)
-    for ans in AnnotationStatus.objects.all()
-)
+annotation_status_choices = [(ans.status, ans.status)
+                             for ans in AnnotationStatus.objects.all()]
 
-annotation_category_choices = tuple(
-    (anc.category, anc.category)
-    for anc in AnnotationCategory.objects.all()
-)
+annotation_category_choices = [(anc.category, anc.category)
+                               for anc in AnnotationCategory.objects.all()]
 
-annotation_type_choices = tuple(
-    (ans.annotation_type, ans.annotation_type)
-    for ans in AnnotationType.objects.all()
-)
+annotation_type_choices = [(ans.annotation_type, ans.annotation_type)
+                           for ans in AnnotationType.objects.all()]
+
 
 class AnnotationForm(forms.Form):
     """Form for editing of annotations."""
@@ -41,7 +35,6 @@ class AnnotationForm(forms.Form):
         choices=annotation_type_choices,
     )
 
-    
     def clean_status(self):
         """Return the Status object."""
         status = self.cleaned_data['status']
@@ -53,10 +46,9 @@ class AnnotationForm(forms.Form):
         category = self.cleaned_data['category']
         obj = AnnotationCategory.objects.get(category=category)
         return obj
-        
+
     def clean_annotation_type(self):
         """Return the Annotationtype object."""
         annotation_type = self.cleaned_data['annotation_type']
         obj = AnnotationType.objects.get(annotation_type=annotation_type)
         return obj
-