+++ lizard_annotation/views.py
@@ -3,18 +3,16 @@
 # Create your views here.
 
 from django.views.generic import FormView
-from django.utils.translation import pgettext 
-from django.utils.translation import ugettext
+from django.utils.translation import pgettext
 from django.views.generic import TemplateView
 from django.views.generic import DetailView
 
 from lizard_annotation.forms import AnnotationForm
 from lizard_annotation.models import Annotation
-from lizard_annotation.models import AnnotationStatus
 from lizard_ui.views import ViewContextMixin
-from lizard_map.views import AppView
 
 from lizard_annotation.mongodb_queries import annotations_list
+
 
 class AnnotationDetailView(ViewContextMixin, DetailView):
 
@@ -37,12 +35,12 @@ class AnnotationEditView(ViewContextMixin, FormView):
         return super(FormView, self).form_valid(form)
 
 
-class AnnotationView(TemplateView):
+class AnnotationView(ViewContextMixin, TemplateView):
 
     template_name = 'lizard_annotation/annotation_view.html'
 
     def annotations(self, reference_filter=""):
-        if reference_filter=="":
+        if reference_filter == "":
             return []
         else:
             return annotations_list(reference_filter)
@@ -55,13 +53,5 @@ class AnnotationView(TemplateView):
         return objs.values()
 
     def post(self, request, *args, **kwargs):
-        reference_filter = request.POST.get('reference_filter')
         context = super(AnnotationView, self).get_context_data(**kwargs)
-        #options = {"annotations": self.annotations(reference_filter)}
         return context
-
-    def get_context_data(self, **kwargs):
-        context = super(AnnotationView, self).get_context_data(**kwargs)
-        return {'annotations': self.annotations(),
-                'reference_objects': self.reference_objects()}
-